A faster hashing implementation with a background thread.
At this point, the only way we're going to get faster is
by doing the commit on a background thread after
handleTransaction has completed. We need to support
FastCopyable in a more robust way to do that, so that
a fast copy of this VMap will see the inflight changes
and not just what is committed to disk, and that the
changes committed to disk will be flushed by the time
the VMap itself is released from Java memory. In this way,
we can employ all 48 cores to help out. We need our
dataSource implementation to be threadsafe to support
this, which will have some impact on performance. It
isn't clear if this will be a win, or by how much, but
we will need to get somebody on it to find out.