Refactor paths to be simple longs.
A change suggested by Atul was to use a simple incrementing
long as the path instead of a byte-packed value, similar to
what would be done with a binary heap representation. In fact,
we tried to do an implementation of VirtualMap that was a
simple b-tree with binary-heap representation, but it was
slower than our b+tree like implementation (long story). I
decided to take the change to the path and apply it to this
branch. Some operations are simpler, some are a little more
cumbersome, but in the end it is a little simpler to
document and explain, so I made the change.