Update HTS Precompile pricing (#2780)
Make gas costs for the HTS precompile to be 120% of the equivalent gas 
cost of the service fee for the related operation. (Network and node fees
are paid by the parent transaction, no need to double-bill).

This required updates to some unit tests as well as the WorkingHours EET.
For the EETs the default payer has an expiry nearly 80 years in the 
future (1 Jan 2100) which causes the gas fee to be in excess of 
11,000,000, which is more than the max gas for a TX.  So we create a new 
user with a reasonable expiry date. Hence, Gas prices demonstrably vary 
with expiry dates.

Signed-off-by: Danno Ferrin <danno.ferrin@hedera.com>