Hardcode range of enums for Seeded SerDes tests
The seeded generation of objects for the large scale serdes test
depends on selecting some constants from a large range of values.
However, when protobuf upgrades occur that range can grow and hence the
values selected for those fields will change.

To address this the default range selector of the objects is fixed to
the 0.25 length. A second form of the call allowing future upgrades to
request values from a larger range is also supplied in the event future
serialization structures depend on the new value.

Signed-off-by: Danno Ferrin <danno.ferrin@hedera.com>