Improve parallelism in HapiApiSpec test clients by creating a common thread pool and avoiding calls to CompletableFuture.runAsync without a thread pool argument. (#3539)
Calls to CompletableFuture.runAsync without a thread pool uses ForkJoinPool.commonPool which is limited to numCores - 1 threads. This is extremely limited and can lead to I/O bounded threads to be starved. This PR addresses this issue by:

This PR:
- Creates a common thread pool with a high upper bound to max threads allowed.
- Switches calls to runAsync to include the new thread pool as the thread pool argument.

On a custom test written, the test previously took 233 seconds to run before this change. After these changes it took 12 seconds to complete.

Signed-off-by: Ray Juang <ray.juang@swirldslabs.com>