Java annotation for interfaces of hedera-evm that are the "interim SPI"
Adds an annotation to identify interfaces/classes that are acceptable for
extending _outside_ of the hedera-evm module. Only interfaces/classes
so identified are extension points.  Thus, use of this annotion declares
the hedera-evm module SPI (service provider interface).

However: This is only an _interim_ SPI.  A clean SPI is needed (will be
designed/provided) in the future.  (And at that time it might be that
no annotation is necessary to mark it, other language features - e.g.
packages - might suffice.

There is no enforcement of this annotation (via an annotation processor)
planned.

Fixes #5115

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>