Add validation (unit) test that all `SystemTask`s are registered in `TaskModule`
System tasks are run by the `SystemTaskManager` which knows about them iff there
is a Dagger-annotated binding for them in `TaskModule`.

But it's easy to forget to add it there.  This validation test ensures that all
defined `SystemTest` implementors are also listed in `TaskModule`.   (This is in
the spirit of the unit tests that check that, at unit test time, that
`bootstrap.properties` defines all known properties and only known properties.)

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>