Cherry pick from 0.37: Traceability migration now gated only on migration feature flag (#6257) (#6272)
Cherry-picked from `release/0.37`:

Traceability migration should run when the migration feature flag is on (and there are contracts which haven't been migrated. But in fact it was also gated on autoRenew entities having been enabled. And too it also needed
to be listed in TaskModule so dagger would include it as a system task to run. Both done here.

And a one-time info-level log message is emitted the first time the traceability task is invoked. After that you can look
for "n-thousand contracts exported" messages (not the exact text) to see progress. (But the current system task
API doesn't allow for a "migration done" message.)

Also adds a validation test (as a unit test) that all `SystemTask`s are registered in `TaskModule`:

- System tasks are run by the `SystemTaskManager` which knows about them iff there
    is a Dagger-annotated binding for them in `TaskModule`.

    But it's easy to forget to add it there.  This validation test ensures that all
    defined `SystemTest` implementors are also listed in `TaskModule`.   (This is in
    the spirit of the unit tests that check that, at unit test time, that
    `bootstrap.properties` defines all known properties and only known properties.)

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>