Move (some) signatures and decoders of versioned ABIs to a dedicated location (#6564)
Start of a new method for having one-stop-shopping for everything we
need related to system contract ABIs.  (Subsequent additions will
include _all_ system contract ABIs, and then moving to use this enum
as the source of ground truth for ABI (names, selectors, etc.)
everywhere.

**N.B.:** The initial goal of this is to simplify correct maintenance by putting everything related to our ABI in one place, instead of being scattered all over the place the way it is now.  The ultimate goal is to _generate_ this enum from the ground truth Solidity system contracts in our smart contracts repo.

Of the versioned ABIs, in this PR are: burn, mint, update token expiry info.

The meat is in the new enum `SystemContractAbis` which provides an enum element for each versioned ABI that supplies things like its selector and its decoder and also other interesting stuff.  All the versioned ABIs are in there, and the unit test checks that the selectors computed from the signatures given match the ones we depend on in `AbiConstants`.

The enum contains _all_ versioned ABIs though this PR only has fixed up the code for burn, mint, and update token expiry.

Future work will finish the work on incorporating the versioned ABIs from this enum into the precompiles, add all the non-versioned ABIs, and eventually refactor with an aim of eliminating other files where this stuff is scattered about, including `ParsingConstants` and `DecodingFacade`.  We should also generate `AbiConstants`.

Subsequent PR will add all the creates, crypto transfer, and update token info.

Fixes #6559

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>