Remove traceability migration code entirely (#7094)
Having been run (in 0.37 and again in 0.38) we no longer need it.  And will no longer need it, ever.

Also it is good to remove it now because migration for contract nonces is coming shortly, and if this code was still in the codebase it might very well be used as a "template". But this code was not implemented appropriately.  The existing `SystemTask` mechanism which was used was _not_ built for migrations, and certainly not this kind of migration.  There were several problems caused during migration attempts that led to operational difficulties and long delay in actually getting the migration to happen (and it took two attempts in production!). A proper migration facility will _not_ be built on `SystemTask`/`SystemTaskManager` (as it stands) and thus should not use this code as a template.

Less code in product **⇒** fewer bugs in production.  What's not to like?

Fixes #6422.  See also #6731.

Signed-off-by: David Bakin <117694041+david-bakin-sl@users.noreply.github.com>