Resolve `Hip17UnhappyTokensSuite` errors (#8614)
This PR fixes all, but one tests in the Hip17UnhappyTokensSuite by:

Adding necessary checks for token deletion in TokenFeeScheduleUpdateHandler, TokenFreezeAccountHandler, TokenMintHandler and TokenUnfreezeAccountHandler
Added 2 new lines in one of the tests, which overrides the configuration for the auto renew duration, as the test create tokens with autoRenewDuration == 100. In most configurations this will fails, as the minimum duration is much higher. This change is mostly to avoid this test failing on most configurations.

Signed-off-by: georgi-l95 <glazarov95@gmail.com>
Signed-off-by: Matt Hess <matt.hess@swirldslabs.com>
Co-authored-by: Matt Hess <matt.hess@swirldslabs.com>