Fix `standardImmutabilitySemanticsHold` in `TokenUpdateSpec` (#9913)
This PR aims to fix the last test in the TokenUpdateSpec, by adding additional validation in the ExpiryValidatorImpl, which is responsible to checking whether the current expiry is smaller number than the new expiry and if it's not, it should throw `INVALID_EXPIRATION_TIME`. This is not valid for Topic, hence the check.

Signed-off-by: georgi-l95 <glazarov95@gmail.com>
Signed-off-by: Venkatachalam Vairavan <vvairava@in.ibm.com>