Migrate contract store from mono-service state to modular-service state
PRELIMINARY FIRST LOOK!  Not tested at all.

Testing will consist of incorporating this into the current `DumpContractStoresCommand` so that it
can either dump the mono-service's store as is, or it can do this migration and then dump the
transformed-into-modular-service's store.  Then the two dumps can be compared - should be identical!

Signed-off-by: David S Bakin <117694041+david-bakin-sl@users.noreply.github.com>