fix: Ethereum contract calls where high-bit of amount is set, fail (#13821)
Ethereum contract calls have an amount - the amount of the native currency to transfer. In Ethereum this amount is unsigned. But as a recent regression we are interpreting this amount as signed (2's complement), leading to failure when the amount's high-order bit is 1.

Bug was introduced in https://github.com/hashgraph/hedera-services/pull/12834.

Fix is to (manually) revert the particular change that treats these amounts as signed.

As a freebie code cleanup also renamed constants `WEIBARS_TO_TINYBAR` to `WEIBARS_IN_A_TINYBAR` to reflect their actual semantic meaning.

Fixes #13816

Signed-off-by: David S Bakin <117694041+david-bakin-sl@users.noreply.github.com>