fix: filter publishing by deactivating tasks
...rather than changing task dependencies.

With this change, we can still call ALL tasks of a certain publishing
profile. E.g, we can do 'releaseMavenCentral' or 'releaseDevelopCommit'
(without leading ':'). This is less risky of breaking if Gradle
is called wrongly. And, more importantly, it allows us to again call
all tasks that publish to CGP repositories (not Nexus) and only exist
in the _platform_ projects.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>