build: define module locations and identities in 'settings.gradle.kts'
This make use of a recently added feature of the
`java-module-dependencies` plugin that offers a notation to
define the directories of modules and their "group:artifact"
ID completely in the 'settings.gradle.kts' file.

With this, the "Module Name" and the "Subproject Names" no longer need
to align. Which caused hard to debug errors in the past when new modules
are added. The plugin now knows about all Module Names by parsing the
'module-info.java' files early.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>