chore: remove 'itest' test set
There are only 4 test classes in this set in the whole code base.
Nothing has been added since more than a year.
Since these were separate, they did not run on CI for several months.

This moves the few tests we have into the default "test" set (the tests
do not take long to run).

There will be the opportunity to sort tests into more categories if
needed in a next step when the current CITR effort (#14407) moves
forward.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>