fix: remove dependencies to 'org.testcontainers' in production code
We only have 'org.testcontainers' in our dependency catalog for testing.
It should never be used in production, as it pulls in the complete test
containers library, with a bunch of transitive dependencies, into the
Hedera Services release.

Here it was mistakenly introduced instead of using the original
'org.bouncycastle' dependencies.

Follow up to #14568

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>