ci: Fix invalid workflow introduced by 15928
**Description**:

PR #15928 moved the jenkins notify steps in
`node-zxc-build-release-artifact.yaml` to its own file,
`node-zxf-deploy-integration.yaml` This file triggers off of the completion of
`node-zxc-build-release-artifact.yaml`. However, the file
`node-flow-deploy-release-artifact.yaml` which calls
build-release-artifact maintained its original calling signature.

This PR removes the `jenkins-integration-url` entry from the
`node-flow-deploy-release-artifact.yaml` file. Which was causing the
workflow to be marked as invalid.

This change should enable the `ZXF: Deploy Production Release` job to
complete.

**Related Issue(s)**:

Relates to: #15947

Signed-off-by: Roger Barker <roger.barker@swirldslabs.com>