Smart contract service metrics, P2 priority.
These are the metrics that are interesting now and we can collect in the consensus node, but later we'll
collect in other ways. E.g., from block nodes. (All these metrics are for post-consensus
transactions and will show up in the record/block stream.)

Does _not_ include metrics for ExchangeRate and PNRG system contracts:  Those system contracts
are coded differently (don't use the same base (abstract) classes as HAS/HSS/HTS) and so will need
a bit more work.

Related issue(s):

Fixes #16088

Reviewer notes:

Changes a boatload of files.  Each of the ~60 FooTranslator classes (about one for each system
contract method) needs a small change, but that's a change to the construtor, and so it had
effects on all the unit tests for those classes as well.  You'll quickly see the pattern ...

Signed-off-by: David S Bakin <117694041+david-bakin-sl@users.noreply.github.com>