ci: Add env token for Check PR Title step to pull-request-formatting workflow.
**Description**:

In commit 778d279 there was a merge for preventing forked repos running on our runners when PRs are created. In this, we removed one env token. This caused an authentication issue in the Check PR Title step. Adding the env token back to the workflow file will resolve the problem.

**Related Issue(s)**:

Fixes #16919

Signed-off-by: Andrew Brandt <andrew.brandt@hashgraph.com>