feat: smart contract service metrics, P2 priority (#16877)
Smart contract service metrics, P2 priority.

These are the metrics that are interesting now and we can collect in the consensus node, but later we'll 
collect in other ways. E.g., from block nodes. (All these metrics are for post-consensus transactions and will show up in the record/block stream.)

Metrics at this time are:
* ERC-20 & ERC-721 call counts
* HAS/HSS/HTS contract call counts, and fail counts
* Direct vs Proxy call counts, and fail counts
* "Category" call counts, and fail counts

Fixes #16088 
Fixes #16868

Signed-off-by: David S Bakin <117694041+david-bakin-sl@users.noreply.github.com>