build: consistently use protobuf '4.29.3'
While we recently updated the runtime to the latest version, the
compiler – protoc – stayed on a lower version. The versions should
be aligned to the latest release. Which is what this PR is doing.

Note that these version changes are not critical as long as the tests
work. Google protobuf is not used in production. It is here to test
that the PBJ generated code is compatible with Google protobuf.

Historically, some utility code lives in the production code and that
is why Google protobuf is currently on the production classpath,
although it is not used in production.

Once this update is done, I would like to get back to #14026, which
is the attempt to move it out of production. We should re-evaluate
how how much work it is now to finally do it.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>